Support for infering the result type of queries in EntityManager::createQuery()
Adds a type parameter "TResult" on the Doctrine\ORM\Query and
Doctrine\ORM\AbstractQuery stubs. This parameter represents the type of the
results returned by the getResult() and execute() family of methods, when in
HYDRATE_OBJECT mode.

Also adds a DynamicReturnTypeExtension on EntityManager::createQuery() so that
TResult is infered from the query string.

Caveat: As the hydration mode influences the return type of the getResult() and
execute() family of methods, we can not just declare that these methods return
TResult yet. This will require a DynamicReturnTypeExtension.